refactor(reservation-app): web mvc 어댑터 모듈 이동
