--- reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/book/infrastructure/web/request/ReservationCreateRequest.java
@@ -1,14 +1,0 @@
-package com.seatliberator.seatliberator.reservation.book.infrastructure.web.request;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-
-import java.time.Instant;
-
-@Schema(description = "예약 생성 요청")
-public record ReservationCreateRequest(
-        @Schema(description = "예약 시작 시간", example = "2026-01-01T13:00Z")
-        Instant startAt,
-        @Schema(description = "예약 종료 시간", example = "2026-01-01T14:30Z")
-        Instant endAt
-) {
-}
