--- reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/room/infrastructure/web/request/SeatCreateRequest.java
@@ -1,10 +1,0 @@
-package com.seatliberator.seatliberator.reservation.room.infrastructure.web.request;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-
-@Schema(description = "좌석 생성 요청")
-public record SeatCreateRequest(
-        @Schema(description = "좌석 ID", example = "A1")
-        String seatId
-) {
-}
