+++ reservation/reservation-web-mvc/src/main/java/com/seatliberator/seatliberator/reservation/infrastructure/web/room/request/CreateRoomRequest.java
@@ -1,0 +1,15 @@
+package com.seatliberator.seatliberator.reservation.infrastructure.web.room.request;
+
+import com.seatliberator.seatliberator.reservation.room.application.port.in.command.CreateRoomCommand;
+import io.swagger.v3.oas.annotations.media.Schema;
+import jakarta.validation.constraints.NotBlank;
+
+@Schema(description = "방 생성 요청")
+public record CreateRoomRequest(
+        @Schema(description = "방 ID", example = "study-room-1")
+        @NotBlank String roomId
+) {
+    public CreateRoomCommand toCommand() {
+        return new CreateRoomCommand(roomId);
+    }
+}
